+++ packages/react/src/diagrams/CardNode/index.tsx
@@ -1,0 +1,13 @@
+import CardNode from './CardNode';
+import { CardNodeColumn } from './CardNodeColumn';
+import { CardNodeLabel } from './CardNodeLabel';
+import { CardNodeSubtitle } from './CardNodeSubtitle';
+import { CardNodeTitle } from './CardNodeTitle';
+
+export {
+	CardNode as default,
+	CardNodeColumn,
+	CardNodeSubtitle,
+	CardNodeTitle,
+	CardNodeLabel,
+};